refactor: move modals folder into components/ui
