+++ src/components/ui/Modals/DeleteResourceModal/index.ts
